moved files to new testcore folder.
