+++ espresso/src/main/java/com/progressive/kherkin/espresso/testcore/PathSegment.kt
@@ -1,4 +1,4 @@
-package com.progressive.kherkin.espresso.custom
+package com.progressive.kherkin.espresso.testcore
 
 import com.progressive.kherkin.common.screen.Screen
